+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 Development:
+  - "validator depositdata" now defaults to mainnet, does not silently fetch fork version from chain
+  - update deposit data output to version 3, to allow for better deposit checking
   - use go-eth2-client for beacon node communications
   - deprecated "--basedir" in favor of "--base-dir"
   - deprecated "--storepassphrase" in favor of "--store-passphrase"